+++ src/IntelliTect.Coalesce.Tests/TargetClasses/TestDbContext/Company.cs
@@ -3,6 +3,7 @@ using IntelliTect.Coalesce.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using IntelliTect.Coalesce.Helpers;
 using System.ComponentModel.DataAnnotations;
+using System;
 
 namespace IntelliTect.Coalesce.Tests.TargetClasses.TestDbContext
 {
@@ -27,7 +28,8 @@ namespace IntelliTect.Coalesce.Tests.TargetClasses.TestDbContext
         public string WebsiteUrl { get; set; }
 
         [DataType(DataType.ImageUrl)]
-        public string LogoUrl { get; set; }
+        [Search(SearchMethod = SearchAttribute.SearchMethods.BeginsWith)]
+        public Uri LogoUrl { get; set; }
 
         [InverseProperty("Company")]
         public ICollection<Person> Employees { get; set; }