+++ src/IntelliTect.Coalesce/TypeDefinition/ParameterViewModel.cs
@@ -70,8 +70,6 @@ namespace IntelliTect.Coalesce.TypeDefinition
 
         public string PascalCaseName => Name.ToPascalCase();
 
-        public bool ConvertsFromJsString => Type.IsNumber || Type.IsString || Type.IsGuid || Type.IsDate || Type.IsBool || Type.IsEnum;
-
         public string CsParameterName => Name.ToCamelCase();
 
         public abstract bool HasDefaultValue { get; }